refactor: move generic state utility functions into state-utils.ts
