+++ packages/server-wallet/src/wallet/store.ts
@@ -10,9 +10,6 @@ import {
   Participant,
   makeDestination,
   deserializeObjective,
-  wireStateToNitroState,
-  deserializeOutcome,
-  convertToInternalParticipant,
   isOpenChannel,
   SignedState,
   objectiveId,
@@ -28,15 +25,14 @@ import {Logger} from 'pino';
 
 import {Channel, ChannelError, CHANNEL_COLUMNS} from '../models/channel';
 import {SigningWallet} from '../models/signing-wallet';
-import {addHash} from '../state-utils';
+import {addHash, addState, clearOldStates, fastDeserializeState} from '../state-utils';
 import {ChannelState, ChainServiceApi} from '../protocols/state';
 import {WalletError, Values} from '../errors/wallet-error';
 import {Bytes32, Address, Uint256, Bytes} from '../type-aliases';
 import {timerFactory, recordFunctionMetrics, setupDBMetrics} from '../metrics';
-import {pick} from '../utilities/helpers';
+import {isReverseSorted, pick} from '../utilities/helpers';
 import {Funding} from '../models/funding';
 import {Nonce} from '../models/nonce';
-import {recoverAddress} from '../utilities/signatures';
 import {ObjectiveModel, DBObjective} from '../models/objective';
 import {AppBytecode} from '../models/app-bytecode';
 import {LedgerRequest, LedgerRequestType} from '../models/ledger-request';
@@ -681,101 +677,4 @@ function ensureReverseSortedOrderOfStates(states: SignedStateVarsWithHash[]): vo
 function validateInvariants(states: SignedStateVarsWithHash[], mySigningAddress: string): void {
   ensureSingleSignedState(states, mySigningAddress);
   ensureReverseSortedOrderOfStates(states);
-}
-
-function isReverseSorted(arr: number[]): boolean {
-  const len = arr.length - 1;
-  for (let i = 0; i < len; ++i) {
-    if (arr[i] < arr[i + 1]) {
-      return false;
-    }
-  }
-  return true;
-}
-
-/**
- * Fold a new signed state into an existing list of signed states (variable parts only).
- *
- * @param vars - list of signed states
- * @param signedState - new signed state
- *
- * A signed state _may_ be a new signature on an existing signed state, in which case
- * we want to "fold" it in by adding a signature to the existing signed state object.
- *
- * In the other case, it represents a brand new state, so we simply push it to the end
- * of the list of existing signed states.
- */
-function addState(
-  vars: SignedStateVarsWithHash[],
-  signedState: SignedStateWithHash
-): SignedStateVarsWithHash[] {
-  const {stateHash, signatures} = signedState;
-
-  const ret = _.cloneDeep(vars);
-
-  const existing = _.find(ret, ['stateHash', stateHash]);
-
-  if (existing) {
-    existing.signatures = _.uniqBy([...existing.signatures, ...signatures], 'signature');
-  } else {
-    ret.push(signedState);
-  }
-
-  return ret;
-}
-
-/**
- * Retrieve a list of states that must be stored to ensure future secure protocol execution.
- *
- * @param signedStates - list of signed states of arbitrary length
- * @param support - a support (list of signed states)
- *
- * Should remove states in signedStates which are not a part of the support.
- */
-function clearOldStates(
-  signedStates: SignedStateVarsWithHash[],
-  support: SignedStateWithHash[] | undefined
-): SignedStateVarsWithHash[] {
-  let sorted = _.reverse(_.sortBy(signedStates, s => s.turnNum));
-
-  // If we don't have a supported state we don't clean anything out
-  if (support && support.length > 0) {
-    // The support is returned in descending turn number so we need to grab
-    // the last element to find the earliest state
-    const {stateHash: firstSupportStateHash} = support[support.length - 1];
-
-    // Find where the first support state is in our current state array
-    const supportIndex = sorted.findIndex(sv => sv.stateHash === firstSupportStateHash);
-
-    // Take everything before that
-    sorted = sorted.slice(0, supportIndex + 1);
-  }
-
-  return sorted;
-}
-
-/**
- * Deserializes a state but uses the wasm utility method local to the server-wallet
- * package. This, as opposed to the JS implementation inside wallet-core.
- */
-function fastDeserializeState(channelId: Bytes32, state: WireSignedState): SignedState {
-  const {outcome, participants, signatures} = state;
-
-  const nitroState = wireStateToNitroState(state);
-
-  const signatureEntries = signatures.map(signature => {
-    const signer = recoverAddress(signature, nitroState);
-
-    if (!_.includes(nitroState.channel.participants, signer))
-      throw new Error(`Recovered address ${signer} is not a participant in channel ${channelId}`);
-
-    return {signature, signer};
-  });
-
-  return {
-    ...state,
-    outcome: deserializeOutcome(outcome),
-    participants: participants.map(convertToInternalParticipant),
-    signatures: signatureEntries,
-  };
 }
